+++ src/main/java/com/hanwul/kbscbackend/domain/account/Account.java
@@ -1,4 +1,4 @@
-package com.hanwul.kbscbackend.domain.user;
+package com.hanwul.kbscbackend.domain.account;
 
 import com.hanwul.kbscbackend.domain.answer.Answer;
 import com.hanwul.kbscbackend.domain.chatting_room.ChattingRoom;
@@ -6,14 +6,15 @@ import com.hanwul.kbscbackend.domain.diary.Diary;
 import com.hanwul.kbscbackend.domain.diarylike.DiaryLike;
 import lombok.Getter;
 import lombok.Setter;
+
 import javax.persistence.*;
 import java.util.ArrayList;
 import java.util.List;
 
 @Entity
 @Getter @Setter
-@Table(name = "user")
-public class User {
+@Table(name = "account")
+public class Account {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
@@ -31,12 +32,12 @@ public class User {
     @OneToMany(mappedBy = "constructor", cascade = CascadeType.ALL)
     private List<ChattingRoom> chattingRoomList = new ArrayList<>();
 
-    @OneToMany(mappedBy = "user", cascade = CascadeType.ALL)
+    @OneToMany(mappedBy = "account", cascade = CascadeType.ALL)
     private List<DiaryLike> diaryLikeList = new ArrayList<>();
 
-    @OneToMany(mappedBy = "user", cascade = CascadeType.ALL)
+    @OneToMany(mappedBy = "account", cascade = CascadeType.ALL)
     private List<Diary> diaryList = new ArrayList<>();
 
-    @OneToMany(mappedBy = "user", cascade = CascadeType.ALL)
+    @OneToMany(mappedBy = "account", cascade = CascadeType.ALL)
     private List<Answer> answerList = new ArrayList<>() ;
 }
