+++ src/main/java/com/hanwul/kbscbackend/domain/account/AccountRepository.java
@@ -1,0 +1,10 @@
+package com.hanwul.kbscbackend.domain.account;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.Optional;
+
+
+public interface AccountRepository extends JpaRepository<Account, Long> {
+    Optional<Account> findByUsername(String username);
+}
