+++ src/main/java/com/hanwul/kbscbackend/domain/security/CustomUserDetailService.java
@@ -1,7 +1,7 @@
 package com.hanwul.kbscbackend.domain.security;
 
-import com.hanwul.kbscbackend.domain.user.User;
-import com.hanwul.kbscbackend.domain.user.UserRepository;
+import com.hanwul.kbscbackend.domain.account.Account;
+import com.hanwul.kbscbackend.domain.account.AccountRepository;
 import lombok.RequiredArgsConstructor;
 
 import org.springframework.security.core.userdetails.UserDetails;
@@ -9,17 +9,21 @@ import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.stereotype.Service;
 
+import java.util.Optional;
+
 @Service
 @RequiredArgsConstructor
 public class CustomUserDetailService implements UserDetailsService {
-    private final UserRepository userRepository;
+    private final AccountRepository userRepository;
     @Override
     public UserDetails loadUserByUsername(String nickname) throws UsernameNotFoundException {
 
-       User user = userRepository.findByNickname(nickname);
+       Optional<Account> result = userRepository.findByUsername(nickname);
 
-        if(user == null) throw new UsernameNotFoundException(String.format("NICKNAME : [%s]을 찾을 수 없습니다", nickname));
+        if(result.isEmpty()) {
+            throw new UsernameNotFoundException(String.format("USERNAME : [%s]을 찾을 수 없습니다", nickname));
+        }
 
-        return new UserSecurityAdapter(user);
+        return new UserSecurityAdapter(result.get());
     }
 }
