--- src/main/java/com/hanwul/kbscbackend/domain/user/UserRepository.java
@@ -1,7 +1,0 @@
-package com.hanwul.kbscbackend.domain.user;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface UserRepository extends JpaRepository<User, Long> {
-
-}
